requirements.txt
# Google Gemini API
google-genai>=0.3.0

# MCP (Model Context Protocol)
mcp>=0.1.0

# Web Scraping
aiohttp>=3.9.1
beautifulsoup4>=4.12.0
html2text>=2020.1.16

# 環境変数管理
python-dotenv>=1.0.0

# テスト用
pytest>=7.4.0
pytest-asyncio>=0.21.0

# 注意: 以下のパッケージはPython標準ライブラリに含まれているため、インストール不要です
# - logging (ロギング)
# - pathlib (パス操作)
# - fcntl (ファイルロック、Unix系OSのみ)
# - json (JSON処理)
# - asyncio (非同期処理)
# - subprocess (プロセス実行)
# - tempfile (一時ファイル)
